+++ src/systems/a320_systems_wasm/src/autobrakes.rs
@@ -1,0 +1,39 @@
+use std::error::Error;
+use std::time::Duration;
+use systems_wasm::aspects::{EventToVariableMapping, EventToVariableOptions, MsfsAspectBuilder};
+use systems_wasm::Variable;
+
+pub(super) fn autobrakes(builder: &mut MsfsAspectBuilder) -> Result<(), Box<dyn Error>> {
+    let options = |options: EventToVariableOptions| {
+        options
+            .ignore_repeats_for(Duration::from_millis(1500))
+            .after_tick_set_to(0.)
+    };
+
+    builder.event_to_variable(
+        "AUTOBRAKE_LO_SET",
+        EventToVariableMapping::Value(1.),
+        Variable::Named("OVHD_AUTOBRK_LOW_ON_IS_PRESSED".into()),
+        options,
+    )?;
+    builder.event_to_variable(
+        "AUTOBRAKE_MED_SET",
+        EventToVariableMapping::Value(1.),
+        Variable::Named("OVHD_AUTOBRK_MED_ON_IS_PRESSED".into()),
+        options,
+    )?;
+    builder.event_to_variable(
+        "AUTOBRAKE_HI_SET",
+        EventToVariableMapping::Value(1.),
+        Variable::Named("OVHD_AUTOBRK_MAX_ON_IS_PRESSED".into()),
+        options,
+    )?;
+    builder.event_to_variable(
+        "AUTOBRAKE_DISARM",
+        EventToVariableMapping::Value(1.),
+        Variable::Named("AUTOBRAKE_DISARM".into()),
+        |options| options.after_tick_set_to(0.),
+    )?;
+
+    Ok(())
+}
